Module check fix, AzureAD check fix and general module code cleanup
found an issue with AzureAD check running install-module MSOnline instead of AzureAD if the version was not correct.
Found an issue with the -match condition for identifying the correct Workload to check, changed that to -eq so that it no longer matches AzureADPreview to AzureAD checks.

did format sanitation for some of the module code.